Totaly seperating REST from the app
