+++ authorization/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+from django.contrib.auth import get_user_model
+
+User = get_user_model()
+
+
+admin.site.register(User)
