+++ boilerplate/urls.py
@@ -7,7 +7,7 @@ from graphene_django.views import GraphQLView
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/v1.0/boilerplate_apps/', include("boilerplate_app.urls", namespace="boilerplate_app-api")),
+    path('boilerplate_apps/', include("boilerplate_app.urls")),
     path('api/', include("rest.urls")),
     path("graphql/", GraphQLView.as_view(graphiql=True)),
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
