+++ boilerplate_app/admin.py
@@ -1,9 +1,7 @@
 from django.contrib import admin
 
-from boilerplate_app.models import User, Projects
+from boilerplate_app.models import Projects
 
 
 # Registering the models
-admin.site.register(User)
 admin.site.register(Projects)
-
