+++ boilerplate_app/models.py
@@ -1,76 +1,7 @@
 from django.db import models
-from django.utils import timezone
-from django.utils.translation import gettext_lazy as _
-from django.contrib.auth.models import (PermissionsMixin, BaseUserManager, AbstractBaseUser)
+from django.contrib.auth import get_user_model
 
-
-class UserManager(BaseUserManager):
-
-    def _create_user(self, email, password, is_active, is_staff, is_superuser, **extra_fields):
-        now = timezone.now()
-        user = self.model(
-            email=email,
-            is_staff=is_staff,
-            is_active=is_active,
-            is_superuser=is_superuser,
-            last_login=now,
-            **extra_fields
-        )
-        user.set_password(password)
-        user.save(using=self._db)
-        return user
-
-    def create_user(self, email, password=None, **extra_fields):
-        return self._create_user(email, password, True, False, False, **extra_fields)
-
-    def create_superuser(self, email, password, **extra_fields):
-        user = self._create_user(email, password, True, True, True, **extra_fields)
-        return user
-
-    def get_queryset(self):
-        return super(UserManager, self).get_queryset()
-
-
-class AutoDateTimeField(models.DateTimeField):
-    def pre_save(self, model_instance, add):
-        return timezone.now()
-
-
-class User(AbstractBaseUser, PermissionsMixin):
-    username = models.CharField(max_length=80, blank=True)
-    email = models.EmailField(_('email address'), unique=True)
-    first_name = models.CharField(_('first name'), max_length=30, blank=True)
-    last_name = models.CharField(_('last name'), max_length=30, blank=True)
-    is_active = models.BooleanField(default=True)
-    is_staff = models.BooleanField(default=False)
-    is_superuser = models.BooleanField(default=False)
-    role = models.CharField(max_length=20, default="admin")
-    created = models.DateTimeField(default=timezone.now)
-    updated = AutoDateTimeField(default=timezone.now)
-
-    objects = UserManager()
-
-    USERNAME_FIELD = 'email'
-
-    class Meta:
-        verbose_name = _('user')
-        verbose_name_plural = _('users')
-
-    def get_full_name(self):
-        """
-        Returns the first_name plus the last_name, with a space in between.
-        """
-        full_name = '%s %s' % (self.first_name, self.last_name)
-        return full_name.strip()
-
-    def get_short_name(self):
-        """
-        Returns the short name for the user.
-        """
-        return self.first_name
-
-    def __str__(self):
-        return self.email
+User = get_user_model()
 
 
 class Projects(models.Model):
